src/third_party/blink/web_tests/external/wpt/css/css-logical/animation-002-expected.txt

This is a testharness.js-based test.
FAIL Logical properties can be animated assert_equals: expected "50px" but got "100px"
FAIL Logical properties in animations respect the writing-mode assert_equals: expected "50px" but got "100px"
FAIL Logical properties in animations respect the direction assert_equals: expected "50px" but got "100px"
PASS Declaration order is respected within @keyframes declaration blocks
PASS Logical properties are able to override physical properties in @keyframes declaration blocks
PASS Declaration order is respected amongst logical properties within @keyframes declaration blocks
FAIL Physical properties and logical properties can be mixed assert_equals: expected "250px" but got "300px"
FAIL Declaration order is respected on each keyframe individually assert_equals: expected "250px" but got "200px"
FAIL Animations update when the writing-mode is changed assert_equals: expected "50px" but got "100px"
PASS Filling animations update when the writing-mode is changed
FAIL The number of interpolating properties can be increased when the writing-mode is changed assert_equals: expected "100px" but got "150px"
FAIL The number of interpolating properties can be decreased when the writing-mode is changed assert_equals: expected "150px" but got "200px"
FAIL Animations update when the writing-mode is changed through a CSS variable assert_equals: expected "50px" but got "100px"
FAIL Animations update when the direction is changed assert_equals: expected "50px" but got "100px"
Harness: the test ran to completion.
